tests moved to routing
