+++ src/Microsoft.DotNet.Interactive.Tests/Parsing/SubmissionParserTests.cs
@@ -368,74 +368,6 @@ var b = 22;";
         }
 
         [Fact]
-        public async Task proxyKernel_does_not_perform_split_if_all_parts_go_to_same_targetKernel_as_the_original_command()
-        {
-            
-            var commandAndEventReceiver = new BlockingCommandAndEventReceiver();
-            var commandAndEventSender = new RecordingKernelCommandAndEventSender();
-
-            commandAndEventSender.OnSend(commandOrEvent =>
-            {
-                if (commandOrEvent.Command is { } command)
-                {
-                    commandAndEventReceiver.Write(new CommandOrEvent( new CommandSucceeded(command)));
-                }
-            });
-            var proxyKernel = new ProxyKernel(
-                "proxyKernel",
-                commandAndEventReceiver,
-                commandAndEventSender);
-
-            var _ = proxyKernel.RunAsync();
-            var code = @"#i ""nuget:source1""
-#i ""nuget:source2""
-#r ""nuget:package1""
-#r ""nuget:package2""
-
-Console.WriteLine(1);";
-            
-            var command = new SubmitCode(code, proxyKernel.Name);
-            await proxyKernel.SendAsync(command);
-
-            commandAndEventSender.Commands.Should().BeEquivalentTo(command);
-        }
-
-
-        [Fact]
-        public async Task proxyKernel_does_not_perform_split_if_all_parts_go_to_same_targetKernel_original_command_has_not_target_kernel()
-        {
-
-            var commandAndEventReceiver = new BlockingCommandAndEventReceiver();
-            var commandAndEventSender = new RecordingKernelCommandAndEventSender();
-
-            commandAndEventSender.OnSend(commandOrEvent =>
-            {
-                if (commandOrEvent.Command is { } command)
-                {
-                    commandAndEventReceiver.Write(new CommandOrEvent(new CommandSucceeded(command)));
-                }
-            });
-            var proxyKernel = new ProxyKernel(
-                "proxyKernel",
-                commandAndEventReceiver,
-                commandAndEventSender);
-
-            var _ = proxyKernel.RunAsync();
-
-            var code = @"#i ""nuget:source1""
-#i ""nuget:source2""
-#r ""nuget:package1""
-#r ""nuget:package2""
-
-Console.WriteLine(1);";
-
-            var command = new SubmitCode(code);
-            await proxyKernel.SendAsync(command);
-
-            commandAndEventSender.Commands.Should().BeEquivalentTo(command);
-        }
-
-        [Fact]
         public void When_targeting_a_local_kernel_after_targeting_a_proxy_kernel_splitting_resumes()
         {
